+++ app/Services/TransactionCategory/TransactionCategoryService.php
@@ -31,6 +31,28 @@ class TransactionCategoryService
 
     }
 
+    public function getByType($type_id) {
+
+        $transCategories = TransactionCategory::where('transaction_type_id', $type_id)->get();
+
+        if (count($transCategories) == 0) 
+            return [
+                "error" => false,
+                "code" => 200,
+                "message" => "No hay categorías registradas",
+                "data" => $transCategories
+            ];
+
+
+        return [
+            "error" => false,
+            "code" => 200,
+            "message" => "Categorías obtenidas con éxito",
+            "data" => $transCategories
+        ];
+
+    }
+
     public function getTransactionCategory($id) {
 
         $transCategory = TransactionCategory::find($id);
@@ -55,27 +77,36 @@ class TransactionCategoryService
 
         $month = $data['month'];
         $year = $data['year'];
+        $type_id = $data['type_id'];
 
         $categories = DB::select(
-            "SELECT 
-                        c.id, 
-                        c.name,
-                        COUNT(t.id) AS total_transactions,
-                        CAST(COALESCE(SUM(t.amount),0) AS SIGNED) AS sum_transactions,
-                        col.hex AS color,
-                        DATE_FORMAT(c.created_at, '%Y-%m-%d') AS created_at,
-                        DATE_FORMAT(c.updated_at, '%Y-%m-%d') AS updated_at
-                    FROM transaction_categories c
-                    INNER JOIN transactions t ON t.transaction_category_id = c.id
-                    INNER JOIN transaction_types tt ON c.transaction_type_id = tt.id
-                    INNER JOIN colors col ON tt.color_id = col.id
-                    WHERE t.user_id = ?
-                    AND MONTH(t.created_at) = ?
-                    AND YEAR(t.created_at) = ?
-                    GROUP BY c.id, c.name, col.hex, c.created_at, c.updated_at
-                    ORDER BY sum_transactions DESC",
-            [$user_id, $month, $year]
-        );
+                        "SELECT 
+                            c.id, 
+                            c.name,
+                            COUNT(t.id) AS total_transactions,
+                            CAST(COALESCE(SUM(t.amount),0) AS SIGNED) AS sum_transactions,
+                            col.hex AS color,
+                            i.icon AS icon,
+                            DATE_FORMAT(c.created_at, '%Y-%m-%d') AS created_at,
+                            DATE_FORMAT(c.updated_at, '%Y-%m-%d') AS updated_at
+                        FROM transaction_categories c
+                        INNER JOIN transactions t 
+                            ON t.transaction_category_id = c.id
+                        INNER JOIN transaction_types tt 
+                            ON c.transaction_type_id = tt.id
+                            AND tt.id = ?
+                        INNER JOIN colors col 
+                            ON tt.color_id = col.id
+                        LEFT JOIN icons i
+                            ON c.icon_id = i.id
+                        WHERE t.user_id = ?
+                            AND MONTH(t.created_at) = ?
+                            AND YEAR(t.created_at) = ?
+                        GROUP BY c.id, c.name, col.hex, i.icon, c.created_at, c.updated_at
+                        ORDER BY sum_transactions DESC",
+                        [$type_id, $user_id, $month, $year]
+                    );
+
 
         if (count($categories) == 0) 
             return [
@@ -84,12 +115,13 @@ class TransactionCategoryService
                 "message" => "No hay categorías registradas",
                 "data" => $categories
             ];
-        
+
         $categories = collect($categories)->map(function($category) {
             $category->total_transactions = (int) $category->total_transactions;
             $category->sum_transactions = (float) $category->sum_transactions;
             return $category;
         });
+
 
         return [
             "error" => false,
@@ -105,6 +137,7 @@ class TransactionCategoryService
         $transCategory = TransactionCategory::create([
             'name' => $data['name'],
             'transaction_type_id' => $data['transaction_type_id'],
+            'icon_id' => $data['icon_id']
         ]);
 
         return [
@@ -126,7 +159,7 @@ class TransactionCategoryService
                 "message" => "Esta categoría no existe",
             ];
 
-        $transCategory->update(Arr::only($data, ['name', 'transaction_type_id']));
+        $transCategory->update(Arr::only($data, ['name', 'transaction_type_id', 'icon_id']));
 
         return [
             "error" => false,
@@ -166,6 +199,14 @@ class TransactionCategoryService
                 "code" => 404,
                 "message" => "Esta categoría no existe",
             ];
+
+        if ($transCategory->transactions()->exists()) {
+            return [
+                "error" => true,
+                "code" => 409,
+                "message" => "No se puede eliminar la categoría porque tiene movimientos relacionados",
+            ];
+        }
 
         $transCategory->delete();
