+++ app/Services/TransactionType/TransactionTypeService.php
@@ -2,6 +2,7 @@
 
 namespace App\Services\TransactionType;
 
+use App\Models\Color;
 use App\Models\TransactionType;
 use Illuminate\Support\Arr;
 
@@ -10,6 +11,15 @@ class TransactionTypeService {
     public static function getAll() {
 
         $transTypes = TransactionType::all();
+
+
+        // $data = $transTypes->map(function ($type) {
+        //     return [
+        //         'id'    => $type->id,
+        //         'name'  => $type->name,
+        //         'color'   => $type->color ? $type->color->hex : null, // hexadecimal del color
+        //     ];
+        // });
 
         if (count($transTypes) == 0) 
             return [
@@ -27,6 +37,43 @@ class TransactionTypeService {
             "data" => $transTypes
         ];
 
+    }
+
+    public function getWithGoal() {
+
+        $transTypes = TransactionType::with('color')->get();
+
+        if (count($transTypes) == 0) 
+            return [
+                "error" => false,
+                "code" => 200,
+                "message" => "No hay tipos de movimientos registrados",
+                "data" => $transTypes
+            ];
+
+        $data = $transTypes->map(function ($type) {
+            return [
+                'id'    => $type->id,
+                'name'  => $type->name,
+                'color'   => $type->color ? $type->color->hex : null,
+            ];
+        });
+
+        
+        $metasColor = Color::where('name', 'Metas')->first();
+
+        $data->push([
+            'id'    => 0, // id arbitrario
+            'name'  => 'Metas',
+            'color' => $metasColor ? $metasColor->hex : null,
+        ]);
+
+        return [
+            "error" => false,
+            "code" => 200,
+            "message" => "Tipos de movimientos obtenidos con éxito",
+            "data" => $data
+        ];
     }
 
     public function getTransactionType($id) {
@@ -54,6 +101,7 @@ class TransactionTypeService {
         $transType = TransactionType::create([
             'name' => $data['name'],
             'color_id' => $data['color_id'],
+            'icon_id' => $data['icon_id'],
         ]);
 
         return [
@@ -75,7 +123,7 @@ class TransactionTypeService {
                 "message" => "Este tipo de movimiento no existe",
             ];
 
-        $transType->update(Arr::only($data, ['name']));
+        $transType->update(Arr::only($data, ['name', 'color_id', 'icon_id']));
 
         return [
             "error" => false,
@@ -116,6 +164,14 @@ class TransactionTypeService {
                 "message" => "Este tipo de movimiento no existe",
             ];
 
+        if ($transType->transactionCategories()->exists()) {
+            return [
+                "error" => true,
+                "code" => 409,
+                "message" => "No se puede eliminar el tipo porque tiene categorias relacionadas",
+            ];
+        }
+            
         $transType->delete();
 
         return [
